Rename yolo loss
